+++ src/db/aio_database.rs
@@ -1,12 +1,8 @@
 use bevy_reflect::Struct;
-use libsql::ffi::Error;
-use libsql::Connection;
 use libsql::Builder;
-use libsql::Database;
+use libsql::Connection;
 use log::debug;
 use log::info;
-use r2d2::ManageConnection;
-use r2d2::Pool;
 use serde::Deserialize;
 use serde::Serialize;
 
@@ -14,13 +10,11 @@ use crate::db::internal::helpers::get_system_char_delimiter;
 use crate::db::internal::queries::alter_table_drop_column;
 use crate::db::internal::queries::alter_table_new_column;
 
-use super::_WalMode;
 use super::aio_query::AnyCountResult;
 use super::aio_query::QueryBuilder;
 use super::aio_query::QueryRowResult;
 use super::aio_query::QueryRowsResult;
 use super::internal::helpers::get_schema_from_generic;
-use super::internal::queries::_set_wal_mode;
 use super::internal::queries::all_query;
 use super::internal::queries::any_count_query;
 use super::internal::queries::change_db_settings;
@@ -35,7 +29,6 @@ use super::internal::queries::get_many_values;
 use super::internal::queries::get_single_value;
 use super::internal::queries::insert_value;
 use super::internal::queries::partial_update;
-use super::internal::queries::set_wal_mode_to_rollback;
 use super::internal::queries::update_value;
 use super::models::Schema;
 
@@ -107,52 +100,26 @@ use super::models::Schema;
 /// ```
 pub struct AioDatabase {
      name: String,
-     conn: Pool<AioDatabaseConnection>,
+     conn: AioDatabaseConnection,
      schema: Box<Vec<Schema>>,
      retries: u32
 }
 
 pub(crate) struct AioDatabaseConnection {
-     builder: Database
-}
-
-impl ManageConnection for AioDatabaseConnection {
-    type Connection = Connection;
-    type Error = Error;
-
-     fn connect(&self) -> Result<Self::Connection, Self::Error> {
-          let result = self.builder.connect();
-          if let Ok(conn) = result {
-               return Ok(conn);
-          }
-          else {
-               return Err(Error::new(-1));
-          }
-     }
-
-     fn is_valid(&self, _conn: &mut Self::Connection) -> Result<(), Self::Error> {
-          return Ok(());
-     }
-
-     fn has_broken(&self, _conn: &mut Self::Connection) -> bool {
-          false
-     }
+     connection: Connection
 }
 
 impl AioDatabase {
      /// Create a locally persisted database. Recommended to run `set_wal_mode` to WAL2 mode after creation.
-     pub async fn create<'a, T>(location: String, name: String, max_pool_size: u32) -> AioDatabase  where T: Default + Struct + Clone  {       
+     pub async fn create<'a, T>(location: String, name: String) -> AioDatabase  where T: Default + Struct + Clone  {       
           let db_location = format!("{}{}{}{}", location, get_system_char_delimiter(), name, ".db");
           let builder = Builder::new_local(db_location).build().await.unwrap();
           let aio_conn = AioDatabaseConnection {
-               builder: builder
+               connection: builder.connect().unwrap()
           };
 
-          let conn_pool = r2d2::Builder::<AioDatabaseConnection>::new().max_size(max_pool_size).build(aio_conn).unwrap();
-          let connection = conn_pool.clone().get().unwrap();
-
           let generic_schema = get_schema_from_generic::<T>();
-          let current_schema_option = get_current_db_schema(&name, &connection).await;
+          let current_schema_option = get_current_db_schema(&name, &aio_conn.connection).await;
 
           if let Some(current_schema) = current_schema_option {
                debug!("Current Db schema: {:?}", current_schema);
@@ -160,7 +127,7 @@ impl AioDatabase {
                for current in current_schema.iter() {
                     if !generic_schema.iter().any(|x| x.field_name == current.field_name) {
                          info!("Dropping column: {}", current.field_name.as_str());
-                         alter_table_drop_column(&name, current.field_name.as_str(), &connection).await;
+                         alter_table_drop_column(&name, current.field_name.as_str(), &aio_conn.connection).await;
                          continue;
                     }
                }
@@ -168,20 +135,20 @@ impl AioDatabase {
                for generic_field in generic_schema.iter() {
                     if !current_schema.iter().any(|x| x.field_name == generic_field.field_name) {
                          info!("Adding column: {} as {}", generic_field.field_name.as_str(), generic_field.field_type.as_str());
-                         alter_table_new_column(&name, generic_field, &connection).await;
+                         alter_table_new_column(&name, generic_field, &aio_conn.connection).await;
                          continue;
                     }
                }
           }
           else {
                debug!("Creating table {} with schema: {:?}", name, generic_schema);
-               create_table(&generic_schema, &name, &connection).await;
-               change_db_settings(&connection).await;
+               change_db_settings(&aio_conn.connection).await;
+               create_table(&generic_schema, &name, &aio_conn.connection).await;
           }
 
           let db = AioDatabase {
                name: name,
-               conn: conn_pool,
+               conn: aio_conn,
                schema: generic_schema,
                retries: 5
           };
@@ -190,18 +157,14 @@ impl AioDatabase {
      }
 
      /// Create an in-memory database.
-     pub async fn create_in_memory<'a, T:  Default + Struct + Clone>(name: String, max_pool_size: u32) -> AioDatabase {
+     pub async fn create_in_memory<'a, T:  Default + Struct + Clone>(name: String) -> AioDatabase {
           let builder = Builder::new_local(":memory:").build().await.unwrap();
           let aio_conn = AioDatabaseConnection {
-               builder: builder
+               connection: builder.connect().unwrap()
           };
 
-          let conn_pool = r2d2::Builder::<AioDatabaseConnection>::new().max_size(max_pool_size).build(aio_conn).unwrap();
-          
-          let connection = conn_pool.clone().get().unwrap();
-
           let generic_schema = get_schema_from_generic::<T>();
-          let current_schema_option = get_current_db_schema(&name, &connection).await;
+          let current_schema_option = get_current_db_schema(&name, &aio_conn.connection).await;
 
           if let Some(current_schema) = current_schema_option {
                debug!("Current Db schema: {:?}", current_schema);
@@ -209,7 +172,7 @@ impl AioDatabase {
                for current in current_schema.iter() {
                     if !generic_schema.iter().any(|x| x.field_name == current.field_name) {
                          info!("Dropping column: {}", current.field_name.as_str());
-                         alter_table_drop_column(&name, current.field_name.as_str(), &connection).await;
+                         alter_table_drop_column(&name, current.field_name.as_str(), &aio_conn.connection).await;
                          continue;
                     }
                }
@@ -217,85 +180,25 @@ impl AioDatabase {
                for generic_field in generic_schema.iter() {
                     if !current_schema.iter().any(|x| x.field_name == generic_field.field_name) {
                          info!("Adding column: {} as {}", generic_field.field_name.as_str(), generic_field.field_type.as_str());
-                         alter_table_new_column(&name, generic_field, &connection).await;
+                         alter_table_new_column(&name, generic_field, &aio_conn.connection).await;
                          continue;
                     }
                }
           }
           else {
                debug!("Creating table {} with schema: {:?}", name, generic_schema);
-               create_table(&generic_schema, &name, &connection).await;
-               change_db_settings(&connection).await;
+               change_db_settings(&aio_conn.connection).await;
+               create_table(&generic_schema, &name, &aio_conn.connection).await;
           }
           
           let db = AioDatabase {
                name: name,
-               conn: conn_pool,
+               conn: aio_conn,
                schema: generic_schema,
                retries: 5
           };
 
           return db;
-     }
-
-     /// Create remote database.
-     pub async fn create_remote_dont_use_only_for_testing<'a, T>(url: String, auth_token: String, name: String ,max_pool_size: u32) -> AioDatabase  where T: Default + Struct + Clone  {       
-          let builder = Builder::new_remote(url, auth_token).build().await.unwrap();
-          let aio_conn = AioDatabaseConnection {
-               builder: builder
-          };
-
-          let conn_pool = r2d2::Builder::<AioDatabaseConnection>::new().max_size(max_pool_size).build(aio_conn).unwrap();
-          let connection = conn_pool.clone().get().unwrap();
-
-          let generic_schema = get_schema_from_generic::<T>();
-          let current_schema_option = get_current_db_schema(&name, &connection).await;
-
-          if let Some(current_schema) = current_schema_option {
-               debug!("Current Db schema: {:?}", current_schema);
-
-               for current in current_schema.iter() {
-                    if !generic_schema.iter().any(|x| x.field_name == current.field_name) {
-                         info!("Dropping column: {}", current.field_name.as_str());
-                         alter_table_drop_column(&name, current.field_name.as_str(), &connection).await;
-                         continue;
-                    }
-               }
-
-               for generic_field in generic_schema.iter() {
-                    if !current_schema.iter().any(|x| x.field_name == generic_field.field_name) {
-                         info!("Adding column: {} as {}", generic_field.field_name.as_str(), generic_field.field_type.as_str());
-                         alter_table_new_column(&name, generic_field, &connection).await;
-                         continue;
-                    }
-               }
-          }
-          else {
-               debug!("Creating table {} with schema: {:?}", name, generic_schema);
-               create_table(&generic_schema, &name, &connection).await;
-               change_db_settings(&connection).await;
-          }
-
-          let db = AioDatabase {
-               name: name,
-               conn: conn_pool,
-               schema: generic_schema,
-               retries: 5
-          };
-
-          return db;
-     }
-
-     /// Set `journal_mode` between WAL or WAL2. 
-     pub(crate) async fn _set_wal_mode(&self, wal_mode: _WalMode) -> Result<(), String> {
-          let conn = self.conn.clone().get().unwrap();
-          return _set_wal_mode(&conn, wal_mode).await;
-     }
-
-     /// Set `journal_mode` to `delete` in order to be compatible to older SQLite versions or to change the mode to WAL2 from WAL.
-     pub  async fn set_wal_mode_to_rollback(&self) {
-          let conn = self.conn.clone().get().unwrap();
-          _ = set_wal_mode_to_rollback(&conn).await;
      }
 
      /// Sets how many retries should be made if a query fails. The delay between retries is 10ms.
@@ -316,14 +219,12 @@ impl AioDatabase {
      /// If set_synchronous(true) then the PRAGMA synchronous will equal to NORMAL (recommended) or false for PRAGMA synchronous to equal to OFF. 
      /// That way transaction will be allowed to be asynchronous which may increase performance but in case of an accident the DB may be corrupted.
      pub async fn set_synchronous(&self, val: bool) {
-          let conn = self.conn.clone().get().unwrap();
-          change_synchronous_settings(&conn, val).await;
+          change_synchronous_settings(&self.conn.connection, val).await;
      }
 
      /// Inserts a **T** value in the database. Returns if the insertion was successful or not after certain retries.
      pub async fn insert_value<'a, T:  Default + Struct + Clone>(&self, value: &T) -> Result<(), String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = insert_value::<T>(&value, self.get_name(), conn, self.retries, false).await;
+          let result = insert_value::<T>(&value, self.get_name(), &self.conn.connection, self.retries, false).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -336,8 +237,7 @@ impl AioDatabase {
 
      /// Inserts a **T** value in the database concurrently. Returns if the insertion was successful or not after certain retries.
      pub(crate) async fn _insert_value_concurrent<'a, T:  Default + Struct + Clone>(&self, value: &T) -> Result<(), String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = insert_value::<T>(&value, self.get_name(), conn, self.retries, true).await;
+          let result = insert_value::<T>(&value, self.get_name(), &self.conn.connection, self.retries, true).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -358,9 +258,7 @@ impl AioDatabase {
      }
      
      pub(crate) async fn get_single_value<'a, T: Default + Struct + Clone>(&self, query_string: String) -> Option<T> {
-          let conn = self.conn.clone().get().unwrap();
-          
-          if let Some(mut query_result) = QueryRowResult::<T>::new(query_string, conn).await {
+          if let Some(mut query_result) = QueryRowResult::<T>::new(query_string, &self.conn.connection).await {
                get_single_value::<T>(&mut query_result);
                return query_result.value;
           }
@@ -370,9 +268,7 @@ impl AioDatabase {
      }
 
      pub(crate) async fn get_many_values<'a, T: Default + Struct + Clone>(&self, query_string: String) -> Option<Vec<T>> {
-          let conn = self.conn.clone().get().unwrap();
-
-          if let Some(mut query_result) = QueryRowsResult::<T>::new_many(query_string, conn).await {
+          if let Some(mut query_result) = QueryRowsResult::<T>::new_many(query_string, &self.conn.connection).await {
                get_many_values::<T>(&mut query_result).await;
                return query_result.value;
           }
@@ -382,8 +278,7 @@ impl AioDatabase {
      }
 
      pub(crate) async fn update_value<'a, T: Default + Struct + Clone>(&self, value: T, where_query: String) -> Result<u64, String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = update_value::<T>(&value, self.get_name(), &where_query, conn, self.retries, false).await;
+          let result = update_value::<T>(&value, self.get_name(), &where_query, &self.conn.connection, self.retries, false).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -395,8 +290,7 @@ impl AioDatabase {
      }
 
      pub(crate) async fn _update_value_concurrent<'a, T: Default + Struct + Clone>(&self, value: T, where_query: String) -> Result<u64, String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = update_value::<T>(&value, self.get_name(), &where_query, conn, self.retries, true).await;
+          let result = update_value::<T>(&value, self.get_name(), &where_query, &self.conn.connection, self.retries, true).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -408,8 +302,7 @@ impl AioDatabase {
      }
 
      pub(crate) async fn partial_update<'a, T: Default + Struct + Clone>(&self, field_name: String, field_value: String, where_query: String) ->  Result<u64, String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = partial_update::<T>(field_name, field_value, self.get_name(), &where_query, conn, self.retries, false).await;
+          let result = partial_update::<T>(field_name, field_value, self.get_name(), &where_query, &self.conn.connection, self.retries, false).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -421,8 +314,7 @@ impl AioDatabase {
      }
 
      pub(crate) async fn _partial_update_concurrent<'a, T: Default + Struct + Clone>(&self, field_name: String, field_value: String, where_query: String) ->  Result<u64, String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = partial_update::<T>(field_name, field_value, self.get_name(), &where_query, conn, self.retries, true).await;
+          let result = partial_update::<T>(field_name, field_value, self.get_name(), &where_query, &self.conn.connection, self.retries, true).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -434,8 +326,7 @@ impl AioDatabase {
      }
 
      pub(crate) async fn delete_value<'a, T: Default + Struct + Clone>(&self, where_query: String) -> Result<u64, String> {
-          let conn = self.conn.clone().get().unwrap();
-          let result = delete_value::<T>(self.get_name(), &where_query, conn, self.retries).await;
+          let result = delete_value::<T>(self.get_name(), &where_query, &self.conn.connection, self.retries).await;
           if let Ok(result) = result {
                return Ok(result);
           }
@@ -447,10 +338,9 @@ impl AioDatabase {
      }
 
      pub(crate) async fn any<'a, T: Default + Struct + Clone>(&self, where_query: String) -> bool {
-          let conn = self.conn.clone().get().unwrap();
           let query = any_count_query::<T>(self.get_name(), &where_query).await;
           
-          if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(query, conn).await {
+          if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(query, &self.conn.connection).await {
                get_single_value::<AnyCountResult>(&mut query_result);
                if let Some(any_result) = query_result.value {
                     return match any_result.count_total {
@@ -468,10 +358,9 @@ impl AioDatabase {
      }
 
      pub(crate) async fn count<'a, T: Default + Struct + Clone>(&self, where_query: String) -> u64 {
-          let conn = self.conn.clone().get().unwrap();
           let query = any_count_query::<T>(self.get_name(), &where_query).await;
           
-          if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(query, conn).await {
+          if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(query, &self.conn.connection).await {
                get_single_value::<AnyCountResult>(&mut query_result);
                if let Some(any_result) = query_result.value {
                     return any_result.count_total;
@@ -486,12 +375,10 @@ impl AioDatabase {
      }
 
      pub(crate) async fn all<'a, T: Default + Struct + Clone>(&self, where_query: String) -> bool {
-          let conn = self.conn.clone().get().unwrap();
-          let conn2 = self.conn.clone().get().unwrap();
           let all_query = all_query::<T>(self.get_name()).await;
           let any_query = any_count_query::<T>(self.get_name(), &where_query).await;
           
-          if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(all_query, conn).await {
+          if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(all_query, &self.conn.connection).await {
                get_single_value::<AnyCountResult>(&mut query_result);
                if let Some(all_result) = query_result.value.clone() {
                     let all_records = all_result.count_total.clone();
@@ -499,7 +386,7 @@ impl AioDatabase {
                     drop(all_result);
                     drop(query_result);
 
-                    if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(any_query, conn2).await {
+                    if let Some(mut query_result) = QueryRowResult::<AnyCountResult>::new(any_query, &self.conn.connection).await {
                          get_single_value::<AnyCountResult>(&mut query_result);
                          if let Some(any_result) = query_result.value {
 
@@ -527,12 +414,9 @@ impl AioDatabase {
           &self,
           index_name: &str,
           columns: Vec<String>) -> Result<(), String> {
-          let conn = self.conn.clone().get().unwrap();
           let query = create_index::<T>(index_name, &self.name, columns);
           
-          _ = conn.execute(&query, ()).await;
-
-          drop(conn);
+          _ = &self.conn.connection.execute(&query, ()).await;
 
           Ok(())
      }
@@ -542,12 +426,9 @@ impl AioDatabase {
           &self,
           index_name: &str,
           columns: Vec<String>) -> Result<(), String> {
-          let conn = self.conn.clone().get().unwrap();
           let query = create_unique_index::<T>(index_name, &self.name, columns);
           
-          _ = conn.execute(&query, ()).await;
-
-          drop(conn);
+          _ = &self.conn.connection.execute(&query, ()).await;
 
           Ok(())
      }
@@ -556,12 +437,9 @@ impl AioDatabase {
      pub async fn drop_index(
           &self,
           index_name: &str) -> Result<(), String> {
-          let conn = self.conn.clone().get().unwrap();
           let query = drop_index(index_name);
           
-          _ = conn.execute(&query, ()).await;
-
-          drop(conn);
+          _ = &self.conn.connection.execute(&query, ()).await;
 
           Ok(())
      }